backend/requirements.txt
django 
djangorestframework
psycopg2-binary
celery 
boto3 
pycurl